Update preparing-domain.md
It's never good practice to store a password in a script. In mine I always prompt for the admin to enter the password.

Using ConvertTo-SecureString (despite providing the password in the file) while using ISE still prompts for the password to be entered and only creates the first account. Alternately using Read-Host -AsSecureString works better (and properly in ISE).